Update dgemm-blocked.c
This code won't compile because I don't fully understand SIMD intrinsics.  The problem is at line 50, error: argument of type "double *" is incompatible with parameter of type "struct __m256d" and I'm stuck on how to fix this.  So I'll leave this here if anybody knows of a quick fix.  Otherwise, I'll see if I can optimize David's code further.